requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
crispy-tailwind==0.5.0
Django==4.0.1
django-crispy-forms==1.14.0
django-htmx==1.11.0
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
